implemented the second api call for menu items added QueryMenu.java to handle those api calls
added dummy restaurants to test and added log messages to output results in logcat
tested output by changing the restaurant address to a string of the JSON to verify that the correct restaurant was getting updated throughout asynchronous callbacks

details are in the comments
added/changed previous skeleton code in comments to update the process

!IMPORTANT!
may run into issues with transitions to the next fragment.
since the api calls are asynchronous when the transition to the next fragment is called the restaurants/menu items may still be null/not complete so parsing to the listView/gmap may be tricky.